+++ handlers/press-reader-entitlements/src/index.ts
@@ -1,20 +1,13 @@
-import { sortBy } from '@modules/arrayFunctions';
 import { getIfDefined } from '@modules/nullAndUndefined';
-import { getProductCatalogFromApi } from '@modules/product-catalog/api';
-import type {
-	ProductCatalog,
-	ProductKey,
-} from '@modules/product-catalog/productCatalog';
-import { ProductCatalogHelper } from '@modules/product-catalog/productCatalog';
 import type { Stage } from '@modules/stage';
 import type {
 	APIGatewayProxyEvent,
 	APIGatewayProxyResult,
 	Handler,
 } from 'aws-lambda';
-import type { SupporterRatePlanItem } from './dynamo';
-import { getSupporterProductData } from './dynamo';
 import { getIdentityId } from './identity';
+import type { SupporterRatePlanItem } from './supporterProductData';
+import { getLatestSubscription } from './supporterProductData';
 import type { Member } from './xmlBuilder';
 import { buildXml } from './xmlBuilder';
 
@@ -24,7 +17,6 @@ export const handler: Handler = async (
 	event: APIGatewayProxyEvent,
 ): Promise<APIGatewayProxyResult> => {
 	console.log(`Input is ${JSON.stringify(event)}`);
-
 	try {
 		if (event.path === '/user-entitlements' && event.httpMethod === 'GET') {
 			const userId = getIfDefined(
@@ -38,19 +30,12 @@ export const handler: Handler = async (
 				statusCode: 200,
 			};
 		}
-
 		return {
 			body: 'Not found',
 			statusCode: 404,
 		};
 	} catch (error) {
 		console.log('Caught exception with message: ', error);
-		if (error instanceof IdentityError) {
-			return {
-				body: 'User not found',
-				statusCode: 404,
-			};
-		}
 
 		return {
 			body: 'Internal server error',
@@ -58,13 +43,6 @@ export const handler: Handler = async (
 		};
 	}
 };
-
-class IdentityError extends Error {
-	constructor(message: string) {
-		super(message);
-		this.name = 'IdentityError';
-	}
-}
 
 export async function getMemberDetails(
 	stage: Stage,
@@ -100,56 +78,4 @@ function createMember(
 				]
 			: [],
 	};
-}
-
-export async function getLatestSubscription(
-	stage: Stage,
-	identityId: string,
-): Promise<SupporterRatePlanItem | undefined> {
-	const supporterProductDataItems = await getSupporterProductData(
-		stage,
-		identityId,
-	);
-
-	if (supporterProductDataItems) {
-		const productCatalog = await getProductCatalogFromApi(stage);
-
-		return getLatestValidSubscription(
-			productCatalog,
-			supporterProductDataItems,
-		);
-	}
-
-	return undefined;
-}
-
-export function getLatestValidSubscription(
-	productCatalog: ProductCatalog,
-	supporterProductData: SupporterRatePlanItem[],
-): SupporterRatePlanItem | undefined {
-	const validProducts: Array<ProductKey | undefined> = [
-		'DigitalSubscription',
-		'HomeDelivery',
-		'NationalDelivery',
-		'SubscriptionCard',
-		'SupporterPlus',
-		'TierThree',
-		// ToDo: add Patron, currently not in product catalog
-	] as const;
-
-	const productCatalogHelper = new ProductCatalogHelper(productCatalog);
-
-	const validSubscriptions = supporterProductData.filter((item) =>
-		validProducts.includes(
-			productCatalogHelper.findProductDetails(item.productRatePlanId)
-				?.zuoraProduct,
-		),
-	);
-
-	const latestSubscription = sortBy(
-		validSubscriptions,
-		(item) => item.termEndDate,
-	).pop();
-
-	return latestSubscription;
 }
